backend/requirement.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
psycopg2-binary==2.9.3
PyJWT==1.7.1
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
